+++ src/test/resources/edgeCases/createModelNewKey.php
@@ -1,0 +1,10 @@
+<?php
+
+class User extends \Illuminate\Database\Eloquent\Model {
+
+}
+
+User::create([
+    'email' => 'email@email.com',
+    '<caret>',
+]);